Rename metadata module
