+++ bee-ledger/src/workers/pruning/batch.rs
@@ -6,7 +6,8 @@ use std::collections::VecDeque;
 use bee_message::{milestone::Milestone, output::OutputId, payload::milestone::MilestoneIndex, Message, MessageId};
 use bee_storage::access::{Batch, Fetch};
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage, Tangle,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    Tangle,
 };
 use hashbrown::{HashMap, HashSet};
 use ref_cast::RefCast;
