+++ bee-node/bee-node/src/tools/rocksdb.rs
@@ -21,7 +21,7 @@ use bee_storage_rocksdb::{
     storage::{Storage, System},
 };
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 use structopt::StructOpt;
 use thiserror::Error;
