+++ bee-protocol/src/workers/propagator.rs
@@ -6,7 +6,7 @@ use std::{any::TypeId, convert::Infallible};
 use async_trait::async_trait;
 use bee_message::{payload::milestone::MilestoneIndex, MessageId};
 use bee_runtime::{node::Node, shutdown_stream::ShutdownStream, worker::Worker};
-use bee_tangle::{metadata::IndexId, solid_entry_point::SolidEntryPoint, Tangle, TangleWorker};
+use bee_tangle::{message_metadata::IndexId, solid_entry_point::SolidEntryPoint, Tangle, TangleWorker};
 use futures::{future::FutureExt, stream::StreamExt};
 use log::*;
 use ref_cast::RefCast;
