+++ bee-storage/bee-storage-memory/src/access/batch.rs
@@ -16,7 +16,7 @@ use bee_storage::{
     backend::StorageBackend,
 };
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 
 use crate::{storage::Storage, table::TableBatch};
