+++ bee-storage/bee-storage-memory/src/access/insert.rs
@@ -17,7 +17,7 @@ use bee_storage::{
     system::System,
 };
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 
 use crate::storage::Storage;
