+++ bee-storage/bee-storage-memory/src/access/multi_fetch.rs
@@ -8,7 +8,7 @@ use std::{iter::Map, vec::IntoIter};
 use bee_ledger::types::{ConsumedOutput, CreatedOutput, OutputDiff};
 use bee_message::{milestone::Milestone, output::OutputId, payload::milestone::MilestoneIndex, Message, MessageId};
 use bee_storage::{access::MultiFetch, backend::StorageBackend, system::System};
-use bee_tangle::{metadata::MessageMetadata, solid_entry_point::SolidEntryPoint};
+use bee_tangle::{message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint};
 
 use crate::storage::Storage;
