+++ bee-storage/bee-storage-memory/src/storage.rs
@@ -18,7 +18,7 @@ use bee_storage::{
     system::{StorageHealth, StorageVersion, System, SYSTEM_HEALTH_KEY},
 };
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 use thiserror::Error;
