+++ bee-storage/bee-storage-rocksdb/src/access/delete.rs
@@ -11,7 +11,7 @@ use bee_message::{
 };
 use bee_storage::access::Delete;
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 use packable::PackableExt;
