+++ bee-storage/bee-storage-rocksdb/src/access/insert.rs
@@ -14,7 +14,7 @@ use bee_storage::{
     system::System,
 };
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 use packable::PackableExt;
