+++ bee-storage/bee-storage-rocksdb/src/access/iter.rs
@@ -12,7 +12,7 @@ use bee_message::{
 };
 use bee_storage::{access::AsIterator, system::System};
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 use packable::PackableExt;
 use parking_lot::RwLockReadGuard;
