+++ bee-storage/bee-storage-rocksdb/src/access/multi_fetch.rs
@@ -6,7 +6,7 @@ use std::{marker::PhantomData, vec::IntoIter};
 use bee_ledger::types::{ConsumedOutput, CreatedOutput, OutputDiff};
 use bee_message::{milestone::Milestone, output::OutputId, payload::milestone::MilestoneIndex, Message, MessageId};
 use bee_storage::{access::MultiFetch, system::System};
-use bee_tangle::{metadata::MessageMetadata, solid_entry_point::SolidEntryPoint};
+use bee_tangle::{message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint};
 use packable::{Packable, PackableExt};
 use parking_lot::RwLockReadGuard;
