+++ bee-storage/bee-storage-rocksdb/src/access/truncate.rs
@@ -10,7 +10,7 @@ use bee_message::{
 };
 use bee_storage::access::Truncate;
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 
 use crate::{
