+++ bee-storage/bee-storage-sled/src/access/iter.rs
@@ -14,7 +14,7 @@ use bee_message::{
 };
 use bee_storage::{access::AsIterator, backend::StorageBackend, system::System};
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 use packable::PackableExt;
