+++ bee-storage/bee-storage-sled/src/access/truncate.rs
@@ -12,7 +12,7 @@ use bee_message::{
 };
 use bee_storage::{access::Truncate, backend::StorageBackend};
 use bee_tangle::{
-    metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
+    message_metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
 
 use crate::{storage::Storage, trees::*};
